Move ConnectionData to Internal.Types (and added BaseTypes).
