+++ src/Network/Mattermost.hs
@@ -138,6 +138,7 @@ import           Control.Arrow ( left )
 
 import           Network.Mattermost.Exceptions
 import           Network.Mattermost.Util
+import           Network.Mattermost.BaseTypes
 import           Network.Mattermost.Internal.Types
 import           Network.Mattermost.Types
