+++ src/Network/Mattermost/Internal/Types.hs
@@ -7,9 +7,37 @@
 
 module Network.Mattermost.Internal.Types where
 
+import Network.Connection (ConnectionContext)
+import Network.HTTP.Headers (Header, HeaderName(..), mkHeader)
+import Network.Mattermost.BaseTypes
 
 data Token = Token String
   deriving (Read, Show, Eq, Ord)
 
 getTokenString :: Token -> String
 getTokenString (Token s) = s
+
+-- For now we don't support or expose the ability to reuse connections,
+-- but we have this field in case we want to support that in the future.
+-- Doing so will require some modifications to withConnection (and uses).
+-- Note: don't export this until we support connection reuse.
+data AutoClose = No | Yes
+  deriving (Read, Show, Eq, Ord)
+
+-- | We return a list of headers so that we can treat
+-- the headers like a monoid.
+autoCloseToHeader :: AutoClose -> [Header]
+autoCloseToHeader No  = []
+autoCloseToHeader Yes = [mkHeader HdrConnection "Close"]
+
+
+data ConnectionData
+  = ConnectionData
+  { cdHostname      :: Hostname
+  , cdPort          :: Port
+  , cdAutoClose     :: AutoClose
+  , cdConnectionCtx :: ConnectionContext
+  , cdToken         :: Maybe Token
+  , cdLogger        :: Maybe Logger
+  , cdUseTLS        :: Bool
+  }
