+++ src/Network/Mattermost/Lenses.hs
@@ -3,6 +3,7 @@
 
 module Network.Mattermost.Lenses where
 
+import Network.Mattermost.Internal.Types
 import Network.Mattermost.Types
 import Network.Mattermost.WebSocket.Types
 import Network.Mattermost.TH
