+++ src/Network/Mattermost/WebSocket.hs
@@ -33,6 +33,7 @@ import qualified Network.WebSockets as WS
 import           Network.WebSockets.Stream (Stream, makeStream)
 
 import           Network.Mattermost.Util
+import           Network.Mattermost.BaseTypes
 import           Network.Mattermost.Internal.Types
 import           Network.Mattermost.Types
 import           Network.Mattermost.WebSocket.Types
